requirements.txt

Django>=5.2.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
python-decouple
Pillow
requests
pytz
gunicorn
reportlab
kavenegar
python-bidi
arabic-reshaper
google-auth
google-auth-oauthlib
google-auth-httplib2
jdatetime
Pillow>=10.0.0
google-auth>=2.15.0
google-auth-oauthlib>=0.7.0
google-auth-httplib2>=0.1.0
requests>=2.28.0